Remove formatter support (replaced by ESLint --fix)
`standard --fix` uses ESLint's built-in formatting functionality. The
nice thing about this vs. a formatter based on esformatter is that it
doesn't require maintaining a whole separate config file for a
different tool. It comes for free from the existing list of ESLint
rules. It also appears to be a lot more reliable than esformatter, with
ES6 support that's as good as ESLint's.

The esformatter integration was some of the trickiest code in this
package. I'm happy to finally be able to remove it.

This should be released as a new major version.