Document all options
Document `plugins` and `envs`, which can be specified in the project's
`package.json`. Change `global` to `globals` since that's what the code
prefers. Note the availability of the aliases.

Document that `cwd` may be provided to `lintText()` and `lintFiles()`,
and that this can cause additional options to be loaded from the
project's `package.json`. This is useful information for linter plugins
which may call this method.

Document that both the `ignore` and `files` globs are resolved relative
to the current working directory.

Document the full set of options, including aliases, and how the
`parser` option takes preference over the `parser` setting in the
project's `package.json`.