wip! Resolve ignore options in standard-engine, not deglob
`standard-engine` has all the information it needs to resolve these
options. `deglob` shouldn't have to repeat the work of finding the
`package.json`.

This will make it easier to control the `ignore` globs in a custom
options parser, for which see
<https://github.com/Flet/standard-engine/pull/135>.

Requires https://github.com/Flet/deglob/pull/6 to be released and the
minimum version to be bumped.